% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{censusdata}
\alias{censusdata}
\title{Census Data}
\format{
An sf object.
}
\usage{
data(censusdata)
}
\description{
This dataset contains Census data.
}
\examples{
data(censusdata)
plot(censusdata)
}
\keyword{datasets}
