% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc_df_demography.R
\docType{data}
\name{df_demography_population_estimates}
\alias{df_demography_population_estimates}
\title{World population projections}
\format{
A tibble with 3 columns and 258 rows
\describe{
\item{year}{The year}
\item{growth}{The UN WPP scenario (one of \code{low}, \code{medium} or \code{high})}
\item{population}{The population estimate}
}
}
\source{
wpp2019 rstat package
}
\usage{
df_demography_population_estimates
}
\description{
A table with the UN World Population Projections
}
\keyword{datasets}
