% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc_df_demography.R
\docType{data}
\name{df_demography_population_growth}
\alias{df_demography_population_growth}
\title{DF Demography: Population growth}
\format{
A tibble with 2 columns and 66 rows
\describe{
\item{year}{The year}
\item{growth}{The growth rate in percent}
}
}
\source{
\url{https://www.census.gov/data-tools/demo/idb/}
}
\usage{
df_demography_population_growth
}
\description{
A table with population growth.
Data was manually scraped…
}
\keyword{datasets}
