% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc_df_demography.R
\docType{data}
\name{df_demography_population_historical}
\alias{df_demography_population_historical}
\title{DF Demography: Historical world population}
\format{
A tibble with 2 columns and 105 rows
\describe{
  \item{year}{The year}
  \item{growth}{The world population estimate}
}
}
\source{
\url{https://www.census.gov/data/tables/time-series/demo/international-programs/historical-est-worldpop.html}
}
\usage{
df_demography_population_historical
}
\description{
A table with historical world population numbers.
}
\keyword{datasets}
