\name{default_path}

\alias{default_path}

\title{default_path}

\description{
Set or get the default download path
}

\usage{
default_path(path)
}

\arguments{
  \item{path}{character. Path name where the data should be downloaded to. If missing, the current default path is returned} 
}

\value{
character
}

