% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_api_center_for_reproductive_rights.R
\name{gdt_crr_get_data}
\alias{gdt_crr_get_data}
\title{Get abortion data from CRR
Get the data for every country from the Center
for Reproductive Rights.}
\usage{
gdt_crr_get_data()
}
\value{
A tibble
}
\description{
Get abortion data from CRR
Get the data for every country from the Center
for Reproductive Rights.
}
\examples{
gdt_crr_get_data() -> reproductive_rights

reproductive_rights |>
  head(5)

}
