% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_api_guttmacher.R
\name{gdt_gutt_get_abortion_rates}
\alias{gdt_gutt_get_abortion_rates}
\title{Get abortion rate from Guttmacher API}
\usage{
gdt_gutt_get_abortion_rates()
}
\value{
A tibble
}
\description{
Get abortion rate from Guttmacher API
}
