% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc_demography.R
\docType{data}
\name{gdt_hmd_demography}
\alias{gdt_hmd_demography}
\title{HMD Demography indicators}
\format{
A tibble with 5 columns and 4729 rows
\describe{
\item{Code}{The HMD country code}
\item{Year}{The year}
\item{Population}{The population}
\item{CBR}{Crude Birth Rate}
\item{CDR}{Crude Death Rate}
}
}
\source{
\url{https://www.mortality.org/}
}
\usage{
gdt_hmd_demography
}
\description{
A table with demography indicators
for all HDM countries (see \code{geotools::gtl_hmd_codes()})
}
\keyword{datasets}
