% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_api_hmd.R
\name{gdt_hmd_lex}
\alias{gdt_hmd_lex}
\title{Get life expectancy from the Human Mortality Database}
\usage{
gdt_hmd_lex(countries, age, type = c("male", "female", "total"))
}
\arguments{
\item{countries}{A vector of country names}

\item{age}{The life expectancy age}

\item{type}{The type (one of \code{male}, \code{female} or \code{total})}
}
\value{
A tibble with 3 columns: \code{year}, \code{country} and \code{lex}
}
\description{
Get life expectancy from the Human Mortality Database
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
# This example isn't run because it
# would need HMD credentials
c("Switzerland", "France") |>
  gdt_hmd_lex(age = 65, type = "total") |>
  head(5)
\dontshow{\}) # examplesIf}
}
