% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_demography.R
\name{gdt_hmd_lex_clean}
\alias{gdt_hmd_lex_clean}
\title{Get clean tibble for LEX data}
\usage{
gdt_hmd_lex_clean(country, age, type)
}
\arguments{
\item{country}{The country name}

\item{age}{The life expectancy age}

\item{type}{The type (one of \code{male}, \code{female} or \code{total})}
}
\value{
A tibble
}
\description{
Get clean tibble for LEX data
}
\keyword{internal}
