% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_api_idb.R
\name{gdt_idb_pyramid}
\alias{gdt_idb_pyramid}
\title{One year population pyramid data (tidy)}
\usage{
gdt_idb_pyramid(country, year)
}
\arguments{
\item{country}{A string with the country name}

\item{year}{An integer with the year}
}
\value{
A dataframe with age, gender, and population
}
\description{
One year population pyramid data (tidy)
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
# Not run: needn a valid IDB API key

gdt_idb_pyramid("Switzerland", 2020)
\dontshow{\}) # examplesIf}
}
