% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_api_idb.R
\name{gdt_idb_pyramid_1y}
\alias{gdt_idb_pyramid_1y}
\title{One year population pyramid data (legacy function)}
\usage{
gdt_idb_pyramid_1y(country, year)
}
\arguments{
\item{country}{A string with the country name}

\item{year}{An integer with the year}
}
\value{
A dataframe with age, female, and male population
}
\description{
One year population pyramid data (legacy function)
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
# Not run: needn a valid IDB API key

gdt_idb_pyramid_1y("Switzerland", 2020)
\dontshow{\}) # examplesIf}
}
