% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_api_idb.R
\name{gdt_idb_tidy_5y_pyramid_data}
\alias{gdt_idb_tidy_5y_pyramid_data}
\title{Cleanup 5y pyramid data}
\usage{
gdt_idb_tidy_5y_pyramid_data(df)
}
\arguments{
\item{df}{A get_idb dataframe}
}
\value{
A tidy dataframe
}
\description{
Cleanup 5y pyramid data
}
\keyword{internal}
