% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_geopolitics.R
\name{gdt_nato_membership}
\alias{gdt_nato_membership}
\title{Table with the NATO membership
states by year and extensions}
\usage{
gdt_nato_membership()
}
\value{
A tibble
}
\description{
Table with the NATO membership
states by year and extensions
}
\examples{
gdt_nato_membership()
}
