% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_api_noaa.R
\name{gdt_ncdc_city_list}
\alias{gdt_ncdc_city_list}
\title{Read NCDC city list from cachedir}
\usage{
gdt_ncdc_city_list()
}
\value{
A dataframe of NCDC locations (cities)
}
\description{
Read NCDC city list from cachedir
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
# Not run: needs local data
ncdc_city_list <- gdt_ncdc_city_list()
\dontshow{\}) # examplesIf}
}
