% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_api_noaa.R
\name{gdt_noaa_climate_data}
\alias{gdt_noaa_climate_data}
\title{Retrieve climate data}
\usage{
gdt_noaa_climate_data(
  location_id,
  data_type = c("temperature", "precipitation")
)
}
\arguments{
\item{location_id}{The ID for the weather station.}

\item{data_type}{The type of data needed. Must be one of \emph{temperature}
or \emph{precipitation}.}
}
\value{
A vector with climate data (precipitation or temperature) for each month
}
\description{
Retrieve climate data from the rnoaa::ncdc NOAA API and compute averages over the
time period 1990-2019.
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
# Not run: needs a valid NOAA API key
gdt_noaa_climate_data("CITY:SZ000003")
\dontshow{\}) # examplesIf}
}
