% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_api_ofs.R
\name{gdt_ofs_vote_list}
\alias{gdt_ofs_vote_list}
\title{Get the swiss vote list}
\usage{
gdt_ofs_vote_list()
}
\value{
A tibble
}
\description{
This function fetches the list of votes in Switzerland
from \href{https://opendata.swiss/}{OpenData.swiss}
}
\examples{
gdt_ofs_vote_list()
}
