% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_api_swissgeo.R
\name{gdt_opendata_swiss_geodata_json}
\alias{gdt_opendata_swiss_geodata_json}
\title{Get swiss geodata in geojson format}
\usage{
gdt_opendata_swiss_geodata_json(...)
}
\value{
A geojson map
}
\description{
Get swiss geodata in geojson format
}
