% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_demography.R
\name{gdt_table_demography}
\alias{gdt_table_demography}
\title{Get general demographic informations}
\usage{
gdt_table_demography(country, years)
}
\arguments{
\item{country}{A country code}

\item{years}{The year for which the information is needed}
}
\value{
A tibble
}
\description{
Get general demographic informations
}
