% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc_demography.R
\docType{data}
\name{gdt_un_hdr_composite}
\alias{gdt_un_hdr_composite}
\title{HMD Demography indicators}
\format{
A tibble with 7 columns and 206824 rows
\describe{
\item{iso3}{The ISO3 country code}
\item{country}{The country}
\item{hdicode}{The 2022 HDI quartile}
\item{indicator}{The HDR indicator}
\item{value}{The value}
}
}
\source{
\url{https://www.mortality.org/}
}
\usage{
gdt_un_hdr_composite
}
\description{
A table with demography indicators
for all HDM countries (see \code{geotools::gtl_hmd_codes()})
}
\keyword{datasets}
