% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_api_noaa.R
\name{gdt_update_ncdc_city_list}
\alias{gdt_update_ncdc_city_list}
\title{Update NCDC city list}
\usage{
gdt_update_ncdc_city_list()
}
\description{
This updates the list of cities available for climate data, and computes the
corresponding country column and store the data locally.
}
\examples{
\dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
# Not run: computationally intensive
gdt_update_ncdc_city_list()
\dontshow{\}) # examplesIf}
}
