% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_api_wb.R
\name{gdt_wb_demo}
\alias{gdt_wb_demo}
\title{Demography data}
\usage{
gdt_wb_demo(country)
}
\arguments{
\item{country}{A string with the country name or iso code}
}
\value{
A tibble with the requested data:
\describe{
\item{population}{Total population}
\item{birth_rate}{Crude Birth Rate (per mil)}
\item{death_rate}{Crude Death Rate (per mil)}
}
}
\description{
This function downloads the data for \emph{population}, \emph{crude birth rate} and \emph{crude death rate}
}
\examples{
gdt_wb_demo("Switzerland")

}
