% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_api_wb.R
\name{gdt_wb_lex}
\alias{gdt_wb_lex}
\title{Life expectancy data}
\usage{
gdt_wb_lex(country)
}
\arguments{
\item{country}{A string with the country name or iso code}
}
\value{
A tibble with the requested data:
\describe{
\item{lex}{Total life expectancy}
\item{lex_male}{Male life expectancy}
\item{lex_female}{Female life expectancy}
}
}
\description{
This function downloads the data for \emph{total life expectancy}, \emph{male life expectancy} and \emph{female life expectancy}
}
