% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geodata-package.R
\docType{package}
\name{geodata-package}
\alias{geodata}
\alias{geodata-package}
\title{geodata: Provide Data for Geography Vizualisations}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Provide a repository of data for geography vizualisations.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://geodata.nenuial.org}
}

}
\author{
\strong{Maintainer}: Pascal Burkhard \email{pascal.burkhard@gmail.com} (\href{https://orcid.org/0009-0001-5504-5084}{ORCID})

}
\keyword{internal}
