% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc_oc_chine.R
\docType{data}
\name{oc_chine_2010_gender_imbalance_by_age}
\alias{oc_chine_2010_gender_imbalance_by_age}
\title{Population chinoise 2018 par sexe et régions}
\format{
Un tibble avec 3 colonnes et 202 lignes
\describe{
\item{age}{Age de la population}
\item{region}{Region \code{urban} ou \code{rural}}
\item{genderbalance}{Nombre de femme moins le nombre d'homme}
}
}
\source{
\url{http://www.stats.gov.cn/tjsj/ndsj/2019/html/E0211.jpg}
}
\usage{
oc_chine_2010_gender_imbalance_by_age
}
\description{
Un tableau de la population des régions
de Chine par sexe pour 2018.
}
\keyword{datasets}
