% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc_oc_chine.R
\docType{data}
\name{oc_chine_2018_sex_ratio_par_region}
\alias{oc_chine_2018_sex_ratio_par_region}
\title{OC Chine: Population 2018 par sexe et régions}
\format{
Un tibble avec 6 colonnes et 31 lignes
\describe{
\item{adm1_code}{Identifiant ADM1 (pour NaturalEarth)}
\item{region}{Nom de la région}
\item{population}{La population totale}
\item{male}{La population féminine}
\item{female}{La population masculine}
\item{ratio}{Le ratio}
}
}
\source{
\url{http://www.stats.gov.cn/tjsj/ndsj/2019/html/E0211.jpg}
}
\usage{
oc_chine_2018_sex_ratio_par_region
}
\description{
Un tableau de la population des régions
de Chine par sexe pour 2018.
}
\keyword{datasets}
