% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc_oc_ecocitoyennete.R
\docType{data}
\name{oc_ecocitoyennete_energy_consumption_per_source}
\alias{oc_ecocitoyennete_energy_consumption_per_source}
\title{OC Ecocitoyenneté: Consommation par source (kWh par personne)}
\format{
Un tibble avec 5 colonnes et 38952 lignes
\describe{
\item{entity}{Pays}
\item{code}{Code ISO3c}
\item{source}{Source d'énergie}
\item{amount}{Quantité d'énergie (kWh par personne)}
}
}
\source{
\url{https://ourworldindata.org/grapher/per-capita-energy-stacked}
}
\usage{
oc_ecocitoyennete_energy_consumption_per_source
}
\description{
Un tableau avec la quantité d'énergie consommée par
personne pour tous les pays du monde (y compris des
données mondiales).
}
\keyword{datasets}
