% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc_oc_geo_au_feminin.R
\docType{data}
\name{oc_geo_au_feminin_2019_gfs_violences_sexuelles_actes}
\alias{oc_geo_au_feminin_2019_gfs_violences_sexuelles_actes}
\title{OC Géo au féminin: Violences sexuelles actes}
\format{
Un tibble avec 2 colonnes et 6 lignes
\describe{
\item{Acte}{Disciptif de l'acte}
\item{Pourcentage}{Pourcentage de femmes qui ont répondues subit cet acte}
}
}
\source{
\url{https://cockpit.gfsbern.ch/fr/cockpit/violence-sexuelles-en-suisse/}
}
\usage{
oc_geo_au_feminin_2019_gfs_violences_sexuelles_actes
}
\description{
Un tableau avec les résultats du sondage gfs.bern de
2019 en rapport avec les violences sexuelles. Réponses
concernant différents actes et le taux de femmes
qui ont déjà vécues cette situation.
}
\keyword{datasets}
