% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc_oc_geo_au_feminin.R
\docType{data}
\name{oc_geo_au_feminin_2019_gfs_violences_sexuelles_experiences}
\alias{oc_geo_au_feminin_2019_gfs_violences_sexuelles_experiences}
\title{OC Géo au féminin: Violences sexuelles expériences}
\format{
Un tibble avec 5 colonnes et 10 lignes
\describe{
\item{Expérience}{Disciptif de l'expérience vécue}
\item{oui}{Nombre de femmes qui ont répondues oui}
\item{ambiguë}{Nombre de femmes qui ont répondues que la situation était ambiguë}
\item{non}{Nombre de femmes qui ont répondues non}
\item{non}{Nombre de femmes qui n'ont pas répondues ou qui ne savait pas}
}
}
\source{
\url{https://cockpit.gfsbern.ch/fr/cockpit/violence-sexuelles-en-suisse/}
}
\usage{
oc_geo_au_feminin_2019_gfs_violences_sexuelles_experiences
}
\description{
Un tableau avec les résultats du sondage gfs.bern de
2019 en rapport avec les violences sexuelles. Réponses
concernant différentes expériences et le taux de femmes
qui ont déjà vécues cette situation.
}
\keyword{datasets}
