% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc_oc_geo_au_feminin.R
\docType{data}
\name{oc_geo_au_feminin_2019_gfs_violences_sexuelles_police}
\alias{oc_geo_au_feminin_2019_gfs_violences_sexuelles_police}
\title{OC Géo au féminin: Violences sexuelles police}
\format{
Un tibble avec 4 colonnes et 6 lignes
\describe{
\item{Raison}{La raison}
\item{oui}{Pourcentage de femmes qui ont répondues que c'était une des raisons}
\item{ne sait pas}{Pourcentage de femmes qui ont répondues que c'était une des raisons}
\item{non}{Pourcentage de femmes qui ont répondues que ce n'était pas une des raisons}
}
}
\source{
\url{https://cockpit.gfsbern.ch/fr/cockpit/violence-sexuelles-en-suisse/}
}
\usage{
oc_geo_au_feminin_2019_gfs_violences_sexuelles_police
}
\description{
Un tableau avec les résultats du sondage gfs.bern de
2019 en rapport avec les violences sexuelles. Réponses
concernant les raisons de ne pas contacter la police.
}
\keyword{datasets}
