% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc_oc_geo_au_feminin.R
\docType{data}
\name{oc_geo_au_feminin_2020_fgm_indirect_data}
\alias{oc_geo_au_feminin_2020_fgm_indirect_data}
\title{OC Géo au féminin: Excision données indirectes}
\format{
Un tibble avec 4 colonnes et 31 lignes
\describe{
\item{Country}{Pays}
\item{Girls_undergone_FGM}{Nombre de femmes qui ont subit une excision}
\item{Legal_status}{Statut légal de l'excision}
\item{iso}{Code ISO3c}
}
}
\source{
\url{https://equalitynow.org/fgm-a-global-picture/}
}
\usage{
oc_geo_au_feminin_2020_fgm_indirect_data
}
\description{
Un tableau avec le nombre d'excisions dans des
pays ne fournissant pas de données directes (pays
pas directement concernés par la pratique).
}
\keyword{datasets}
