% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc_oc_geo_au_feminin.R
\docType{data}
\name{oc_geo_au_feminin_2021_ofs_avortements}
\alias{oc_geo_au_feminin_2021_ofs_avortements}
\title{OC Géo au féminin: Avortement en Suisse}
\format{
Un tibble avec 3 colonnes et 34 lignes
\describe{
\item{year}{Année}
\item{name}{Population}
\item{rate}{Taux d'avortement (pour 1000 femmes)}
}
}
\source{
\url{https://www.bfs.admin.ch/bfs/fr/home/statistiques/sante/etat-sante/reproductive/interruptions-grossesses.html}
}
\usage{
oc_geo_au_feminin_2021_ofs_avortements
}
\description{
Un tableau du taux d'avortement en Suisse
pour les adolescentes et les femmes en général
de 2005 à 2020.
}
\keyword{datasets}
