% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc_oc_geo_au_feminin.R
\docType{data}
\name{oc_geo_au_feminin_2021_unhdr_gii}
\alias{oc_geo_au_feminin_2021_unhdr_gii}
\title{OC Géo au féminin: Indicateur GII}
\format{
Un tibble avec 3 colonnes et 170 lignes
\describe{
\item{country}{Pays}
\item{gii}{Indicateur GII}
\item{iso}{Code pays ISO3c}
}
}
\source{
\url{https://hdr.undp.org/data-center/thematic-composite-indices/gender-inequality-index#/indicies/GII}
}
\usage{
oc_geo_au_feminin_2021_unhdr_gii
}
\description{
Un tableau avec l'indicateur GII (Gender Inequality Index)
pour l'année 2021
}
\keyword{datasets}
