% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc_oc_geo_au_feminin.R
\docType{data}
\name{oc_geo_au_feminin_2022_ofs_avortements_par_ages}
\alias{oc_geo_au_feminin_2022_ofs_avortements_par_ages}
\title{OC Géo au féminin: Avortements par canton}
\format{
Un tibble avec 3 colonnes et 18 lignes
\describe{
\item{age_group}{Tranche d'âge}
\item{unit}{Unité (n ou \%)}
\item{abortion}{Mesure de l'avortement}
}
}
\usage{
oc_geo_au_feminin_2022_ofs_avortements_par_ages
}
\description{
Un tableau avec les avortements par tranche d'âge
en Suisse en 2022
}
\keyword{datasets}
