% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc_oc_geo_au_feminin.R
\docType{data}
\name{oc_geo_au_feminin_2022_ofs_avortements_par_canton}
\alias{oc_geo_au_feminin_2022_ofs_avortements_par_canton}
\title{Avortement suisse par canton}
\format{
Un tibble avec 5 colonnes et 494 lignes
\describe{
\item{age_group}{Tranche d'âge}
\item{unit}{Unité (n ou \%)}
\item{region}{Région du pays}
\item{canton}{Le canton}
\item{abortion}{Mesure de l'avortement}
}
}
\source{
\url{https://www.bfs.admin.ch/bfs/fr/home/statistiques/catalogues-banques-donnees/tableaux.assetdetail.26386394.html}
}
\usage{
oc_geo_au_feminin_2022_ofs_avortements_par_canton
}
\description{
Un tableau avec les avortements par canton en Suisse
en 2022
}
\keyword{datasets}
