% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc_oc_geo_au_feminin.R
\docType{data}
\name{oc_geo_au_feminin_2023_oecd_sigi}
\alias{oc_geo_au_feminin_2023_oecd_sigi}
\title{Indicateur SIGI}
\format{
Un tibble avec 8 colonnes et 1635 lignes
\describe{
\item{REGION}{Code région}
\item{Region}{Intitulé région}
\item{LOCATION}{Code pays ISO3c}
\item{Country}{Pays}
\item{VAR}{Code variable}
\item{Variables}{Intitulé variable}
\item{Time}{Année}
\item{Value}{Valeur}
}
}
\source{
\url{https://stats.oecd.org/Index.aspx?DataSetCode=SIGI2023}
}
\usage{
oc_geo_au_feminin_2023_oecd_sigi
}
\description{
Un tableau avec l'indicateur SIGI (Social Institutions and Gender Index)
pour l'année 2023
}
\keyword{datasets}
