% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc_oc_geo_au_feminin.R
\docType{data}
\name{oc_geo_au_feminin_oecd_programmation_16_24_ans}
\alias{oc_geo_au_feminin_oecd_programmation_16_24_ans}
\title{OC Géo au féminin: Pourcentage de jeunes entre 16 et 24 ans qui savent programmer}
\format{
Un tibble avec 5 colonnes et 818 lignes
\describe{
\item{ISO}{Code pays ISO3c}
\item{Country}{Le pays}
\item{Scope}{Hommes, femmes, ou tous}
\item{Time}{L'année}
\item{Value}{La valeur en pourcent}
}
}
\source{
\url{https://goingdigital.oecd.org/indicator/54}
}
\usage{
oc_geo_au_feminin_oecd_programmation_16_24_ans
}
\description{
Un tableau avec le pourcentage de jeunes (16 à 24 ans)  qui savent
programmer pour les différents pays de l'OECD.
}
\keyword{datasets}
