% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc_oc_geo_au_feminin.R
\docType{data}
\name{oc_geo_au_feminin_ofs_avortements_par_region}
\alias{oc_geo_au_feminin_ofs_avortements_par_region}
\title{OC Géo au féminin: Avortements par région}
\format{
Un tibble avec 3 colonnes et 18 lignes
\describe{
\item{annee}{Année}
\item{region}{Région}
\item{avortement}{Taux d'avortement}
}
}
\usage{
oc_geo_au_feminin_ofs_avortements_par_region
}
\description{
Un tableau avec les avortements par régions
en Suisse entre 2007 et 2022
}
\keyword{datasets}
