% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc_oc_geo_au_feminin.R
\docType{data}
\name{oc_geo_au_feminin_ofs_election_femmes}
\alias{oc_geo_au_feminin_ofs_election_femmes}
\title{OC Géo au féminin: Femmes en politique fédérale}
\format{
Un tibble avec 3 colonnes et 38 lignes
\describe{
\item{year}{L'année}
\item{percentage}{Le pourcentage}
\item{council}{L'instance}
}
}
\source{
\url{https://www.bfs.admin.ch/bfs/fr/home/statistiques/politique/elections/femmes-elues.html}
}
\usage{
oc_geo_au_feminin_ofs_election_femmes
}
\description{
Un tableau avec le pourcentage de femmes dans les différentes
instances fédérale (Conseil National, Conseil des États et
Conseil Fédéral) en fonction de l'année
}
\keyword{datasets}
