% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc_oc_geo_au_feminin.R
\docType{data}
\name{oc_geo_au_feminin_ofs_election_femmes_conseil_federal}
\alias{oc_geo_au_feminin_ofs_election_femmes_conseil_federal}
\title{OC Géo au féminin: Femmes au Conseil Fédéral}
\format{
Un tibble avec 2 colonnes et 12 lignes
\describe{
\item{year}{L'année}
\item{percentage}{Le pourcentage}
}
}
\source{
\url{https://www.bfs.admin.ch/bfs/fr/home/statistiques/politique/elections/femmes-elues.html}
}
\usage{
oc_geo_au_feminin_ofs_election_femmes_conseil_federal
}
\description{
Un tableau avec le pourcentage de femmes au Conseil Fédéral
en fonction de l'année
}
\keyword{datasets}
