% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc_oc_geo_au_feminin.R
\docType{data}
\name{oc_geo_au_feminin_ofs_election_femmes_conseil_national}
\alias{oc_geo_au_feminin_ofs_election_femmes_conseil_national}
\title{Femmes au Conseil National}
\format{
Un tibble avec 2 colonnes et 13 lignes
\describe{
\item{year}{L'année}
\item{percentage}{Le pourcentage}
}
}
\source{
\url{https://www.bfs.admin.ch/bfs/fr/home/statistiques/politique/elections/femmes-elues.html}
}
\usage{
oc_geo_au_feminin_ofs_election_femmes_conseil_national
}
\description{
Un tableau avec le pourcentage de femmes au Conseil National
en fonction de l'année
}
\keyword{datasets}
