% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc_oc_geo_au_feminin.R
\docType{data}
\name{oc_geo_au_feminin_owid_suffrage_feminin}
\alias{oc_geo_au_feminin_owid_suffrage_feminin}
\title{Suffrage féminin}
\format{
Un tibble avec 3 colonnes et 206 lignes
\describe{
\item{Entity}{Le pays}
\item{Code}{Code ISO3c}
\item{Year}{Année d'accès au suffrage féminin}
}
}
\source{
\url{https://ourworldindata.org/grapher/universal-suffrage-women-lexical}
}
\usage{
oc_geo_au_feminin_owid_suffrage_feminin
}
\description{
Un tableau avec l'année à laquelle les femmes ont obtenus le droit de vote
dans les différents pays du monde.
}
\keyword{datasets}
