% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc_oc_geo_au_feminin.R
\docType{data}
\name{oc_geo_au_feminin_sex_ratio}
\alias{oc_geo_au_feminin_sex_ratio}
\title{OC Géo au féminin: Sex-ratio à la naissance}
\format{
Un tibble avec 6 colonnes et 43472 lignes
\describe{
\item{UN_Code}{Code UN}
\item{Region}{Type de région UN}
\item{Location}{Nom du pays/région}
\item{Year}{L'année}
\item{SRB}{Sex-ration à la naissance}
\item{iso}{Code pays ISO3c}
}
}
\source{
\url{https://population.un.org/wpp/}
}
\usage{
oc_geo_au_feminin_sex_ratio
}
\description{
Un tableau avec le sex ratio à la naissance tiré des données
WPP 2022
}
\keyword{datasets}
