% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc_oc_geo_au_feminin.R
\docType{data}
\name{oc_geo_au_feminin_votations_avortement}
\alias{oc_geo_au_feminin_votations_avortement}
\title{Votations sur l'avortement}
\format{
Un tibble avec 3 colonnes et 6 lignes
\describe{
\item{Votation}{Le titre de la votation}
\item{Oui}{Pourcentage de voix pour le oui}
\item{Non}{Pourcentage de voix pour le non}
}
}
\usage{
oc_geo_au_feminin_votations_avortement
}
\description{
Un tableau avec les résultats au niveau fédéral
de différentes votations en rapport avec
l'avortement
}
\keyword{datasets}
