% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc_oc_geo_au_feminin.R
\docType{data}
\name{oc_geo_au_feminin_wpp2012_lex}
\alias{oc_geo_au_feminin_wpp2012_lex}
\title{Espérance de vie WPP2012}
\format{
Un tibble avec 5 colonnes et 236 lignes
\describe{
\item{un_code}{Code pays UN}
\item{period}{Période de l'estimation}
\item{lexf}{Espérance de vie des femmes}
\item{lexm}{Espérance de vie des hommes}
\item{lexdelta}{Différence d'espérance de vie}
}
}
\source{
\url{https://population.un.org/wpp/}
}
\usage{
oc_geo_au_feminin_wpp2012_lex
}
\description{
Un tableau de l'espérance de vie selon les données WPP2012
par pays.
}
\keyword{datasets}
