% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc_oc_geo_au_feminin.R
\docType{data}
\name{oc_geo_au_feminin_wpp2019_sex_ratio}
\alias{oc_geo_au_feminin_wpp2019_sex_ratio}
\title{OC Géo au féminin: Rapport des sexes}
\format{
Un tibble avec 7 colonnes et 78435 lignes
\describe{
\item{un_code}{Code pays UN}
\item{name}{Période de l'estimation}
\item{age}{Tranche d'age}
\item{year}{Année}
\item{female}{Population féminine}
\item{male}{Population masculine}
\item{ratio}{Rapport hommes/femmes}
}
}
\source{
\url{https://population.un.org/wpp/}
}
\usage{
oc_geo_au_feminin_wpp2019_sex_ratio
}
\description{
Un tableau du rapport des sexes pour différentes années
selon les chiffres du WPP2019.
}
\keyword{datasets}
