% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc_oc_israel_palestine.R
\docType{data}
\name{oc_israel_palestine_2008_population_nae}
\alias{oc_israel_palestine_2008_population_nae}
\title{Population israélienne née à l'étranger}
\format{
Un tibble avec 3 colonnes et 5 lignes
\describe{
\item{country}{Le pays de naissance}
\item{continent}{Le continent de naissance}
\item{population}{Nombre absolu de personnes}
}
}
\source{
\url{https://www.cbs.gov.il/en/Statistics/Pages/Generators/Generator-of-the-2008-Census.aspx}
}
\usage{
oc_israel_palestine_2008_population_nae
}
\description{
Un tableau avec l'origine pour la population
née à l'étranger en Israël en 2008.
}
\keyword{datasets}
