% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc_oc_israel_palestine.R
\docType{data}
\name{oc_israel_palestine_immigration_israel}
\alias{oc_israel_palestine_immigration_israel}
\title{Immigration en Israël}
\format{
Un tibble avec 2 colonnes et 69 lignes
\describe{
\item{year}{L'année d'immigration}
\item{immigration}{Nombre absolu de personnes}
}
}
\source{
\url{https://www.cbs.gov.il/he/publications/doclib/2018/4.\%20shnatonimmigration/st04_02.pdf}
}
\usage{
oc_israel_palestine_immigration_israel
}
\description{
Un tableau avec l'immigration en Israël par année
en chiffre absolus de 1949 à 2017.
}
\keyword{datasets}
