% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc_oc_israel_palestine.R
\docType{data}
\name{oc_israel_palestine_immigration_juive_avant_1948}
\alias{oc_israel_palestine_immigration_juive_avant_1948}
\title{OC Israël-Palestine: Immigration juive en Palestine avant 1948}
\format{
Un tibble avec 2 colonnes et 6 lignes
\describe{
\item{period}{La période}
\item{immigration}{Le nombre approximatif de Juifs qui immigrent}
}
}
\source{
\url{https://www.bpb.de/themen/migration-integration/laenderprofile/english-version-country-profiles/58400/historical-development-of-jewish-immigration/}
}
\usage{
oc_israel_palestine_immigration_juive_avant_1948
}
\description{
Un tableau avec l'immigration \strong{approximative}
en Palestine avant 1948. Montre les cinq premières
Aliyah juives.
}
\keyword{datasets}
