% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc_oc_russie.R
\docType{data}
\name{oc_russie_2010_elections_ukraine}
\alias{oc_russie_2010_elections_ukraine}
\title{Élection présidentielle d'Ukraine en 2010}
\format{
Un tibble avec 4 colonnes et 27 lignes
\describe{
\item{CODE}{Code adm_1}
\item{REGION}{Nom de la région}
\item{CANDIDATE}{Candidat remportant la région}
\item{PERCENT}{Pourcentage des voix}
}
}
\usage{
oc_russie_2010_elections_ukraine
}
\description{
Un tableau du résultat des élections présidentielle
en Ukraine en 2010.
}
\keyword{datasets}
