% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc_oc_russie.R
\docType{data}
\name{oc_russie_2012_religion}
\alias{oc_russie_2012_religion}
\title{Religion par région}
\format{
Un tibble avec 4 colonnes et 1422 lignes
\describe{
\item{region}{Nom de la région en anglais}
\item{religion}{\describe{
\item{Athee}{Athéistes}
\item{Orthodoxe}{Orthodoxe pratique (membre de l'Eglise)}
\item{Deiste}{Crois en Dieu mais sans appartenance}
\item{Christianisme}{Christianisme}
\item{Hindouisme}{Hindouisme}
\item{Judaisme}{Judaïsme}
\item{Islam}{Islam}
\item{Traditionnel}{Croyances traditionnelles}
\item{Bouddhisme}{Bouddhisme}
\item{Pentecotisme}{Pentecôtisme}
\item{Autre}{Autre}
\item{Pas de réponse}{Pas de réponse}
}}
\item{percent}{Pourcentage}
\item{adm1_code}{Identifiant ADM1 (pour NaturalEarth)}
}
}
\source{
\url{https://sreda.org/arena}
}
\usage{
oc_russie_2012_religion
}
\description{
Un tableau de la religion par région en 2012.
}
\keyword{datasets}
