% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc_oc_russie.R
\docType{data}
\name{oc_russie_2019_esperance_vie}
\alias{oc_russie_2019_esperance_vie}
\title{Espérance de vie en Russie}
\format{
Un tibble avec 5 colonnes et 560 lignes
\describe{
\item{region}{Nom de la région en russe}
\item{lex}{Espérance de vie}
\item{adm1_code}{Identifiant ADM1 (pour NaturalEarth)}
}
}
\source{
\url{https://rosstat.gov.ru/storage/mediabank/IeUTMwMW/prodolzhitelnost.xlsx}
}
\usage{
oc_russie_2019_esperance_vie
}
\description{
Un tableau de l'espérance de vie en Russie
par régions pour 2019.
}
\keyword{datasets}
