% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc_oc_russie.R
\docType{data}
\name{oc_russie_2019_fecondite}
\alias{oc_russie_2019_fecondite}
\title{OC Russie: Indice de fécondité}
\format{
Un tibble avec 4 colonnes et 170 lignes
\describe{
\item{region}{Nom de la région en russe}
\item{adm1_code}{Identifiant ADM1 (pour NaturalEarth)}
\item{year}{Année}
\item{tfr}{Indice de fécondité}
}
}
\source{
\url{https://rosstat.gov.ru/storage/mediabank/7-1(2).xlsx}
}
\usage{
oc_russie_2019_fecondite
}
\description{
Un tableau de l'indice de fécondité en Russie
par régions.
}
\keyword{datasets}
