% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc_oc_russie.R
\docType{data}
\name{oc_russie_2019_migration}
\alias{oc_russie_2019_migration}
\title{Migrations en Russie}
\format{
Un tibble avec 5 colonnes et 2507 lignes
\describe{
\item{adm1_code}{ADM1 ID (for NaturalEarth)}
\item{country}{Country name}
\item{year}{Year of data observation}
\item{type}{One of \emph{arrival} or \emph{departure}}
\item{migration}{Number of people migrating}
}
}
\source{
\url{https://rosstat.gov.ru/storage/mediabank/migr2.xls} with additional Google Translate
}
\usage{
oc_russie_2019_migration
}
\description{
Un tableau des migrations en Russie avec les
immigration et émigration de 1997 à 2019.
}
\keyword{datasets}
