% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc_oc_russie.R
\docType{data}
\name{oc_russie_2019_natalite_mortalite}
\alias{oc_russie_2019_natalite_mortalite}
\title{Mortalité et natalité}
\format{
Un tibble avec 5 colonnes et 174 lignes
\describe{
\item{region}{Nom de la région en russe}
\item{type}{Type de mesure: absolute ou per1000}
\item{indicator}{Indicateur: cbr, cdr ou rni}
\item{data}{Donnée}
\item{adm1_code}{Identifiant ADM1 (pour NaturalEarth)}
}
}
\source{
\url{https://rosstat.gov.ru/storage/mediabank/EDN(1).xlsx}
}
\usage{
oc_russie_2019_natalite_mortalite
}
\description{
Un tableau de l'indice de la natalité et mortalité en Russie
par régions pour 2019.
}
\keyword{datasets}
