% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc_oc_russie.R
\docType{data}
\name{oc_russie_2020_population_composition}
\alias{oc_russie_2020_population_composition}
\title{Religion par région}
\format{
Un tibble avec 4 colonnes et 1422 lignes
\describe{
\item{type_pop}{Population urbaine ou rurale}
\item{sexe}{Homme, femme ou les deux}
\item{ethnie}{Nom de l'ethnie}
\item{ethnie_detail}{Détail concernant de l'ethnie}
\item{population}{Population}
}
}
\source{
\url{https://rosstat.gov.ru/storage/mediabank/Tom5_tab1_VPN-2020.xlsx}
}
\usage{
oc_russie_2020_population_composition
}
\description{
Un tableau de la religion par région en 2012.
}
\keyword{datasets}
