% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc_oc_russie.R
\docType{data}
\name{oc_russie_2020_population_municipale}
\alias{oc_russie_2020_population_municipale}
\title{OC Russie: population municipale en 2020}
\format{
Un tibble avec 4 colonnes et 2539 lignes
\describe{
\item{name}{Nom de la province}
\item{oktmo}{Code de la municipalité}
\item{2016}{Population en 2016}
\item{2020}{Population en 2020}
}
}
\source{
\url{https://showdata.gks.ru/}
}
\usage{
oc_russie_2020_population_municipale
}
\description{
Un tableau de la population municipale en Russie
en 2016 et 2020.
}
\keyword{datasets}
