% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc_oc_russie.R
\docType{data}
\name{oc_russie_2023_accroissement}
\alias{oc_russie_2023_accroissement}
\title{Taux d'accroissement naturel en Russie}
\format{
Un tibble avec 5 colonnes et 591 lignes
\describe{
\item{codes}{OKTMO Codes}
\item{region}{Nom de la région en russe}
\item{adm1_code}{Identifiant ADM1 (pour NaturalEarth)}
\item{year}{Année}
\item{rni}{Taux d'accroissement naturel}
}
}
\source{
\url{https://showdata.gks.ru/report/297732}
}
\usage{
oc_russie_2023_accroissement
}
\description{
Un tableau du taux d'accroissement naturel en Russie
par régions.
}
\keyword{datasets}
