% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc_oc_russie.R
\docType{data}
\name{oc_russie_2023_fecondite}
\alias{oc_russie_2023_fecondite}
\title{Indice de fécondité en Russie}
\format{
Un tibble avec 5 colonnes et 595 lignes
\describe{
\item{codes}{OKTMO Codes}
\item{region}{Nom de la région en russe}
\item{adm1_code}{Identifiant ADM1 (pour NaturalEarth)}
\item{year}{Année}
\item{tfr}{Indice de fécondité}
}
}
\source{
\url{https://showdata.gks.ru/report/278974}
}
\usage{
oc_russie_2023_fecondite
}
\description{
Un tableau de l'indice de fécondité en Russie
par régions.
}
\keyword{datasets}
