% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc_oc_russie.R
\docType{data}
\name{oc_russie_2023_population}
\alias{oc_russie_2023_population}
\title{Population municipale en Russie}
\format{
Un tibble avec 5 colonnes et 2806 lignes
\describe{
\item{codes}{OKTMO Codes}
\item{region}{Nom de la région en russe}
\item{adm1_code}{Identifiant ADM1 (pour NaturalEarth)}
\item{year}{Année}
\item{pop}{Population}
}
}
\source{
\url{https://showdata.gks.ru/report/278930}
}
\usage{
oc_russie_2023_population
}
\description{
Un tableau de la population municipale en Russie
de 1990 à 2023.
}
\keyword{datasets}
