% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc_oc_russie.R
\docType{data}
\name{oc_russie_adm6_gis}
\alias{oc_russie_adm6_gis}
\title{ADM6 GIS}
\format{
Un tibble avec 27 colonnes et 2338 lignes
\describe{
\item{osm_id}{OSM id}
\item{osm_way_id}{OSM way id}
\item{name}{Name}
\item{type}{OSM tourism tag}
\item{aeroway}{OSM aeroway tag}
\item{amenity}{OSM amenity tag}
\item{admin_level}{OSM admin_level}
\item{boundary}{OSM boundary tag}
\item{barrier}{OSM barrier tag}
\item{building}{OSM building tag}
\item{craft}{OSM craft tag}
\item{geological}{OSM geological tag}
\item{land_area}{OSM land_area tag}
\item{landuse}{OSM landuse tag}
\item{leisure}{OSM leisure tag}
\item{man_made}{OSM man_made tag}
\item{military}{OSM military tag}
\item{natural}{OSM natural tag}
\item{office}{OSM office tag}
\item{place}{OSM place tag}
\item{shop}{OSM shop tag}
\item{sport}{OSM sport tag}
\item{tourism}{OSM tourism tag}
\item{other_tags}{OSM other tags}
\item{oktmo}{OKTMO ID}
}
}
\source{
Open Street Map and WikiData
}
\usage{
oc_russie_adm6_gis
}
\description{
Un tableau geographique des 2338 communes russes.
}
\keyword{datasets}
