% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc_oc_russie.R
\docType{data}
\name{oc_russie_avortements_national}
\alias{oc_russie_avortements_national}
\title{OC Russie: avortements national}
\format{
Un tibble avec 2 colonnes et 25 lignes
\describe{
\item{year}{Année}
\item{abortions}{Nombre d'avortements pour 100 naissances}
}
}
\source{
\url{https://showdata.gks.ru/olap2/descr/report/297822/}
}
\usage{
oc_russie_avortements_national
}
\description{
Un tableau des avortements en Russie
de 1992 à 2017 pour 100 naissances.
}
\keyword{datasets}
