% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc_oc_russie.R
\docType{data}
\name{oc_russie_levada_attitude_eu}
\alias{oc_russie_levada_attitude_eu}
\title{OC Russie: Attitude de la Russie vis-à-vis de l'Europe}
\format{
Un tibble avec 3 colonnes et 363 lignes
\describe{
\item{attitude}{Positive, Negative ou Difficile à dire}
\item{date}{Date du sondage}
\item{percentage}{Pourcentage de l'observation}
}
}
\source{
\url{https://www.levada.ru/en/ratings/attitudes-towards-countries/}
}
\usage{
oc_russie_levada_attitude_eu
}
\description{
Un tableau de l'évolution de l'attitude de la population
russe vis-à-vis de l'Union Européenne. Sondage Levada.
}
\keyword{datasets}
