% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc_oc_russie.R
\docType{data}
\name{oc_russie_mariages}
\alias{oc_russie_mariages}
\title{OC Russie: Mariages}
\format{
Un tibble avec 5 colonnes et 1636 lignes
\describe{
\item{region}{Nom de la région en anglais}
\item{oktmo}{OKTMO Codes}
\item{year}{Année}
\item{mariages}{Taux de mariage pour 1000 habitants}
\item{adm1_code}{Identifiant ADM1 (pour NaturalEarth)}
}
}
\source{
\url{https://showdata.gks.ru/olap2/descr/report/297890/}
}
\usage{
oc_russie_mariages
}
\description{
Un tableau des mariages en Russie
par régions de 2000 à 2019.
}
\keyword{datasets}
