% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc_oc_russie.R
\docType{data}
\name{oc_russie_mariages_divorces_national}
\alias{oc_russie_mariages_divorces_national}
\title{OC Russie: Mariages et divorces national}
\format{
Un tibble avec 3 colonnes et 60 lignes
\describe{
  \item{year}{Année}
  \item{indicator}{L'îndicateur (mariages ou divorces)}
  \item{data}{La donnée numérique}
}
}
\source{
\url{https://showdata.gks.ru/olap2/descr/report/297882/}
\url{https://showdata.gks.ru/olap2/descr/report/297880/}
}
\usage{
oc_russie_mariages_divorces_national
}
\description{
Un tableau des mariages et divorces en Russie
de 1990 à 2019.
}
\keyword{datasets}
