% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc_oc_russie.R
\docType{data}
\name{oc_russie_otan}
\alias{oc_russie_otan}
\title{OTAN}
\format{
Un tibble avec 3 colonnes et 33 lignes
\describe{
\item{country}{Pays}
\item{iso3}{Code ISO3c}
\item{year}{Année d'adhésion}
}
}
\source{
\url{https://www.nato.int/cps/en/natohq/topics_52044.htm}
}
\usage{
oc_russie_otan
}
\description{
Un tableau de l'année d'adhésion à l'OTAN
}
\keyword{datasets}
