% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc_oc_russie.R
\docType{data}
\name{oc_russie_suicides}
\alias{oc_russie_suicides}
\title{Suicides en Russie}
\format{
Un tibble avec 5 colonnes et 1858 lignes
\describe{
\item{codes}{OKTMO Codes}
\item{adm1_code}{Identifiant ADM1 (pour NaturalEarth)}
\item{region}{Nom de la région en anglais}
\item{year}{Année}
\item{suicides}{Nombre de suicides}
}
}
\source{
\url{https://showdata.gks.ru/olap2/descr/report/297886/}
}
\usage{
oc_russie_suicides
}
\description{
Un tableau des suicides en Russie
par régions de 1997 à 2019.
}
\keyword{datasets}
