% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc_oc_russie.R
\docType{data}
\name{oc_russie_suicides_national}
\alias{oc_russie_suicides_national}
\title{Suicides en Russie}
\format{
Un tibble avec 2 colonnes et 30 lignes
\describe{
\item{year}{Année}
\item{suicides}{Nombre de suicides}
}
}
\source{
\url{https://showdata.gks.ru/olap2/descr/report/297886/}
}
\usage{
oc_russie_suicides_national
}
\description{
Un tableau des suicides en Russie
par régions de 1997 à 2019.
}
\keyword{datasets}
