% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc_oc_suisse.R
\docType{data}
\name{oc_suisse_1970_04_12_suffrage_feminin_vs}
\alias{oc_suisse_1970_04_12_suffrage_feminin_vs}
\title{Suffrage féminin au Valais}
\format{
Un tibble avec 3 colonnes et 13 lignes
\describe{
\item{id}{Identifiant OFS du district}
\item{vote_oui}{Nombre de vote oui}
\item{vote_non}{Nombre de vote non}
}
}
\source{
\url{https://www.letempsarchives.ch/page/GDL_1970_04_13/9}
}
\usage{
oc_suisse_1970_04_12_suffrage_feminin_vs
}
\description{
Un tableau des résultats par districts du vote concernant le suffrage
féminin au Valais le 12 avril 1970.
}
\keyword{datasets}
