% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc_oc_suisse.R
\docType{data}
\name{oc_suisse_2000_religion}
\alias{oc_suisse_2000_religion}
\title{Religion par commune}
\format{
Un tibble avec 4 colonnes et 55024 lignes
\describe{
\item{ID}{Identifiant de la commune}
\item{Commune}{Nom de la commune}
\item{Religion}{Religion}
\item{Pourcentage}{Pourcentage}
}
}
\source{
\url{https://www.pxweb.bfs.admin.ch/pxweb/fr/px-x-0103010000_151/-/px-x-0103010000_151.px/}
}
\usage{
oc_suisse_2000_religion
}
\description{
Un tableau avec le pourcentage des religions par commune
en Suisse en 2000. Oui c'est vieux, mais visiblement l'OFS
n'a pas de données après le dernier recensement...
}
\keyword{datasets}
