% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc_oc_suisse.R
\docType{data}
\name{oc_suisse_2021_immigration_italienne}
\alias{oc_suisse_2021_immigration_italienne}
\title{Immigration italienne}
\format{
Un tibble avec 3 colonnes et 39 lignes
\describe{
\item{year}{Année d'observation}
\item{population}{Population résidante italienne en Suisse au 1er janvier}
\item{immigration}{Immigration italienne en Suisse}
}
}
\source{
\url{https://www.pxweb.bfs.admin.ch/pxweb/fr/px-x-0103010000_151/-/px-x-0103010000_151.px/}
}
\usage{
oc_suisse_2021_immigration_italienne
}
\description{
Un tableau avec les chiffres de la population et de l'immigration italienne
en Suisse de 1981 à 2019.
}
\keyword{datasets}
