% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc_oc_suisse.R
\docType{data}
\name{oc_suisse_2023_nationalite}
\alias{oc_suisse_2023_nationalite}
\title{Nationalite en Suisse}
\format{
Un tibble avec 2 colonnes et 210 lignes
\describe{
\item{Nationalite}{La nationalité}
\item{Population}{Population (chiffre absolu)}
}
}
\source{
\url{https://www.pxweb.bfs.admin.ch/pxweb/fr/px-x-0103010000_151/-/px-x-0103010000_151.px/}
}
\usage{
oc_suisse_2023_nationalite
}
\description{
Un tableau avec le nombre de personnes par nationalité
en Suisse en 2023.
}
\keyword{datasets}
