% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc_oc_suisse.R
\docType{data}
\name{oc_suisse_suffrage_feminin}
\alias{oc_suisse_suffrage_feminin}
\title{Suffrage féminin en suisse}
\format{
Un tibble avec 2 colonnes et 26 lignes
\describe{
\item{Canton}{Nom du canton}
\item{Annee}{Année de l'obtention du suffrage féminin}
}
}
\source{
\url{https://fr.wikipedia.org/wiki/Suffrage_féminin_en_Suisse}
}
\usage{
oc_suisse_suffrage_feminin
}
\description{
Un tableau des années d'introduction du suffrage féminin en Suisse
par cantons.
}
\keyword{datasets}
