% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc_df_demography.R
\docType{data}
\name{un_wpp_2019_period_estimates}
\alias{un_wpp_2019_period_estimates}
\title{World population period estimates 2019}
\format{
A tibble with 25 columns and 14940 rows
\describe{
\item{LocID}{Location ID}
\item{Location}{Location}
\item{VarID}{Variant ID}
\item{Variant}{Variant}
\item{Time}{Period}
\item{MidPeriod}{Middle of the period}
\item{TFR}{Total fertility (live births per woman)}
\item{NRR}{Net reproduction rate (surviving daughters per woman)}
\item{CBR}{Crude birth rate (births per 1,000 population)}
\item{Births}{Number of births, both sexes combined (thousands)}
\item{LEx}{Life expectancy at birth for both sexes combined (years)}
\item{LExMale}{Male life expectancy at birth (years)}
\item{LExFemale}{Female life expectancy at birth (years)}
\item{IMR}{Infant mortality rate, q(1), for both sexes combined (infant deaths per 1,000 live births)}
\item{Q5}{Under-five mortality, 5q0, for both sexes combined (deaths under age five per 1,000 live births)}
\item{CDR}{Crude death rate (deaths per 1,000 population)}
\item{Deaths}{Number of deaths, both sexes combined (thousands)}
\item{DeathsMale}{Number of male deaths (thousands)}
\item{DeathsFemale}{Number of female deaths (thousands)}
\item{CNMR}{Net migration rate (per 1,000 population)}
\item{NetMigrations}{Net number of migrants, both sexes combined (thousands)}
\item{GrowthRate}{Average annual rate of population change (percentage)}
\item{NatIncr}{Rate of natural increase (per 1,000 population)}
\item{SRB}{Sex ratio at birth (male births per female births)}
\item{MAC}{Female mean age of childbearing (years)}
}
}
\source{
\url{https://population.un.org/wpp2019/Download/Standard/Population/}
}
\usage{
un_wpp_2019_period_estimates
}
\description{
A table with the UN World Population Projections
}
\keyword{datasets}
