% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_api.R
\name{util_chk_api_call}
\alias{util_chk_api_call}
\title{Check API connexion}
\usage{
util_chk_api_call(call)
}
\arguments{
\item{call}{An API url}
}
\value{
Returns an error if the API call fails
}
\description{
Check API connexion
}
\keyword{internal}
