% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_api.R
\name{util_chk_idb_api_key}
\alias{util_chk_idb_api_key}
\title{Check if IDB API key is defined}
\usage{
util_chk_idb_api_key(api_key = NULL)
}
\arguments{
\item{api_key}{}
}
\value{
Returns an error if the IDB API key is not properly set
}
\description{
Check if IDB API key is defined
}
\keyword{internal}
