% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_api.R
\name{util_dwnl_api_json}
\alias{util_dwnl_api_json}
\title{Download API data}
\usage{
util_dwnl_api_json(call)
}
\arguments{
\item{call}{An API url}
}
\value{
Returns API content
}
\description{
Download API data
}
\keyword{internal}
