% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc_df_cosmology.R
\docType{data}
\name{df_cosmology_solar_system_objects}
\alias{df_cosmology_solar_system_objects}
\title{DF Cosmology: Solar System Objects}
\format{
A tibble with 10 columns and 10 rows
\describe{
  \item{english}{The english name of the object}
  \item{french}{The french name of the object}
  \item{type_en}{Type of object in english}
  \item{type_fr}{Type of object in french}
  \item{distance}{The distance to the Sun in AU}
  \item{mass}{The masse in kilogram}
  \item{radius}{The radius in meter}
  \item{density}{The density in kilgram per cubic meters}
  \item{rotation}{The rotation period in hours or days}
  \item{revolution}{The revolution period compared to Earth}
}
}
\source{

}
\usage{
df_cosmology_solar_system_objects
}
\description{
A table with solar system objects.
}
\keyword{datasets}
