% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_demography.R
\name{gdt_hmd_birth_raw}
\alias{gdt_hmd_birth_raw}
\title{HMD raw births data}
\usage{
gdt_hmd_birth_raw(country)
}
\arguments{
\item{country}{Country code}
}
\value{
A dataframe
}
\description{
HMD raw births data
}
