% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_api_hmd.R, R/data_demography.R
\name{gdt_hmd_lex}
\alias{gdt_hmd_lex}
\title{Get life expectancy table from Humand Mortality Database}
\usage{
gdt_hmd_lex(countries, age, type = c("male", "female", "total"))

gdt_hmd_lex(countries, age, type = c("male", "female", "total"))
}
\arguments{
\item{countries}{A list of country names}

\item{age}{The life expectancy age}

\item{type}{The type (one of \code{male}, \code{female} or \code{total})}

\item{country}{A string with the country name}
}
\value{
A tibble

A tibble with 3 columns: \code{year}, \code{country} and \code{lex}
}
\description{
Get life expectancy table from Humand Mortality Database

Get life expectancy from the Human Mortality Database
}
