% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_demography.R
\name{gdt_hmd_lex}
\alias{gdt_hmd_lex}
\title{Get life expectancy from the Human Mortality Database}
\usage{
gdt_hmd_lex(countries, age, type = c("male", "female", "total"))
}
\arguments{
\item{countries}{A list of country names}

\item{age}{The life expectancy age}

\item{type}{The type (one of `male`, `female` or `total`)}
}
\value{
A tibble with 3 columns: `year`, `country` and `lex`
}
\description{
Get life expectancy from the Human Mortality Database
}
