% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_api_hmd.R
\name{gdt_hmd_mortality}
\alias{gdt_hmd_mortality}
\title{Get mortality table from Human Mortality Database}
\usage{
gdt_hmd_mortality(country)
}
\arguments{
\item{country}{A string with the country name}
}
\value{
A demogdata object
}
\description{
Get mortality table from Human Mortality Database
}
