% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_api_idb.R
\name{gdt_idb_pyramid_5y}
\alias{gdt_idb_pyramid_5y}
\title{Return 5y population data}
\usage{
gdt_idb_pyramid_5y(country, year)
}
\arguments{
\item{country}{A string with the country name}

\item{year}{An integer with the year}
}
\value{
A dataframe
}
\description{
Return 5y population data
}
\examples{
gdt_idb_pyramid_5y("Switzerland", 2020)

}
