% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_api_swissgeo.R
\name{gdt_opendata_swiss_geodata_json}
\alias{gdt_opendata_swiss_geodata_json}
\title{Get swiss geodata in geojson format}
\usage{
gdt_opendata_swiss_geodata_json(...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[=gdt_opendata_swiss_geodata]{gdt_opendata_swiss_geodata}}
  \describe{
    \item{\code{geolevel}}{One of "national", "canton", "district", "municipality", "zh_counting_districts" or "lakes"}
    \item{\code{latest}}{Latest geo data (otherwise beginning of the year data)}
    \item{\code{verbose}}{Boolean}
  }}
}
\value{
A geojson map
}
\description{
Get swiss geodata in geojson format
}
\examples{
gdt_opendata_swiss_geodata_json(geolevel = "national")

}
