% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_demography.R
\name{gdt_table_demography}
\alias{gdt_table_demography}
\title{General demographic informations}
\usage{
gdt_table_demography(country, years)
}
\arguments{
\item{country}{A country code}

\item{years}{The year for which the information is needed}
}
\value{
A tibble
}
\description{
General demographic informations
}
\examples{
gdt_table_demography("Switzerland", 2020)

}
