\name{geodata-package}
\alias{geodata-package}
\alias{geodata}
\docType{package}


\title{
\packageTitle{geodata}
}

\description{
\packageDescription{geodata}

\tabular{ll}{
\bold{Function} \tab \bold{Description} \cr
\code{\link{bio_oracle}}   \tab  Marine data from bio-oracle\cr
\code{\link{cmip6_world}}   \tab  Downscaled and calibrated CMIP6 projected future climate data\cr
\code{\link{cmip6_tile}}   \tab  Downscaled and calibrated CMIP6 data by tile\cr
\code{\link{country_codes}}\tab   Country codes \cr
\code{\link{crop_calendar_sacks}}\tab   Sachs crop calendar data\cr
\code{\link{crop_monfreda}}\tab   Monfreda crop data (area, yield)\cr
\code{\link{crop_spam}}    \tab   SPAM crop data (area, yield, value)\cr
\code{\link{cropland}}\tab   Cropland density for the world from three sources\cr
\code{\link{elevation_3s}} \tab   Elevation data for tile (3 seconds resolution) \cr
\code{\link{elevation_30s}} \tab   Elevation data for by country (30 seconds resolution) \cr
\code{\link{elevation_global}} \tab  Global elevation data (various resolutions) \cr
\code{\link{gadm}} \tab   Administrative boundaries for any country in the world  \cr
\code{\link{world}}\tab   Boundaries for the countries in the world  \cr
\code{\link{landcover}} \tab   Global landcover data \cr
\code{\link{footprint}} \tab   Human footprint data \cr
\code{\link{osm}} \tab   OpenStreetMap data by country  \cr
\code{\link{population}}\tab   Download population density data \cr
\code{\link{soil_af}}      \tab  Chemical and physical soil properties data for Africa for different soil depths \cr
\code{\link{soil_af_water}} \tab   Physical soil properties  for Africa for water balance computations \cr
\code{\link{soil_af_elements}} \tab   Soil element concentration data for Africa \cr
\code{\link{soil_af_isda}} \tab   Soil data for Africa derived from the iDSA data set\cr
\code{\link{soil_world_vsi}}     \tab   Virtually connect to the global soilgrids data \cr
\code{\link{soil_world}}   \tab   Global soils data \cr
\code{\link{sp_occurrence}}\tab   Species occurrence data from the Global Biodiversity Information Facility\cr
\code{\link{travel_time}}\tab  Travel time to cities and ports \cr
\code{\link{worldclim_global}}\tab   Global climate data\cr
\code{\link{worldclim_country}}\tab  Climate data by country\cr
\code{\link{worldclim_tile}} \tab   Climate data by tile\cr
}}

