\name{geodata-package}
\alias{geodata-package}
\alias{geodata}
\docType{package}


\title{
\packageTitle{geodata}
}

\description{
\packageDescription{geodata}

\tabular{ll}{
\code{\link{function}}\tab Description \cr
\code{\link{cmip6_world}}   \tab  Downscaled and calibrated CMIP6 projected future climate data\cr
\code{\link{country_codes}}\tab   Country codes \cr
\code{\link{crop_calendar_sacks}}\tab   Sachs crop calendar data\cr
\code{\link{crop_monfreda}}\tab   Monfreda crop data (area, yield)\cr
\code{\link{crop_spam}}    \tab   SPAM crop data (area, yield, value)\cr
\code{\link{elevation_3s}} \tab   Elevation data for tile (3 seconds resolution) \cr
\code{\link{elevation_30s}} \tab   Elevation data for by country (30 seconds resolution) \cr
\code{\link{elevation_global}} \tab  Global elevation data (various resolutions) \cr
\code{\link{gadm}} \tab   Get administrative boundaries for any country in the world  \cr
\code{\link{world}}\tab   Get boundaries for the countries in the world  \cr
\code{\link{cropland_africa}}\tab   Download cropland extent data for Africa \cr
\code{\link{population}}\tab   Download population density data \cr
\code{\link{soil_af_elements}} \tab   Connect to or download chemical soil element concentration (for the 0-30 cm topsoil) data for Africa \cr
\code{\link{soil_af_isda}} \tab   Soil data for Africa derived from the iDSA data set\cr
\code{\link{soil_af}}      \tab   Chemical soil properties data for Africa for different soil depths \cr
\code{\link{soil_world_vsi}}     \tab   Virtually connect to the global soilgrids data \cr
\code{\link{soil_world}}   \tab   Global soils data \cr
\code{\link{sp_occurrence}}\tab   Species occurrence records from the Global Biodiversity Information Facility\cr
\code{\link{worldclim_global}}\tab   Glocal climate data\cr
\code{\link{worldclim_country}}\tab  Climate data by country\cr
\code{\link{worldclim_tile}} \tab   Climate data by tile\cr
}}

