% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc_oc_russie.R
\docType{data}
\name{oc_russie_2019_migration}
\alias{oc_russie_2019_migration}
\title{OC Russie: Évolution de la population}
\format{
Un tibble avec 2 colonnes et 30 lignes
\describe{
\item{adm1_code}{Identifian ADM1 (pour NaturalEarth)}
\item{country}{Country name}
\item{year}{Year of data observation}
\item{type}{One of \code{arrival} or \code{departure}}
\item{migration}{Number of people migrating}
}
}
\source{
\url{https://rosstat.gov.ru/storage/mediabank/migr2.xls} with additional Google Translate
}
\usage{
oc_russie_2019_migration
}
\description{
Un tableau de l'évolution de la population en Russie
par régions à 2019.avec un comparaison 1990 - 2020
}
\keyword{datasets}
