% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc_oc_russie.R
\docType{data}
\name{oc_russie_2020_evolution_population}
\alias{oc_russie_2020_evolution_population}
\title{OC Russie: Évolution de la population}
\format{
Un tibble avec 2 colonnes et 30 lignes
\describe{
\item{region}{Nom de la région en anglais}
\item{oktmo}{OKTMO Code}
\item{adm1_code}{Identifian ADM1 (pour NaturalEarth)}
\item{\code{2020}}{Population en 2020}
\item{\code{1990}}{Population en 1990}
\item{solde}{Pourcentage d'évolution sur ces 30 dernières anneés}
}
}
\source{
\url{https://showdata.gks.ru/olap2/descr/report/278928/}
}
\usage{
oc_russie_2020_evolution_population
}
\description{
Un tableau de l'évolution de la population en Russie
par régions à 2019.avec un comparaison 1990 - 2020
}
\keyword{datasets}
