% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc_oc_russie.R
\docType{data}
\name{oc_russie_depenses_militaires}
\alias{oc_russie_depenses_militaires}
\title{Dépenses militaires}
\format{
Un tibble avec 4 colonnes et 3411 lignes
\describe{
\item{country}{Pays}
\item{year}{Année}
\item{military_expenditure}{Dépense militaire en mio $ US}
\item{iso3c}{Identifiant ISO3c}
}
}
\source{
\url{https://milex.sipri.org/sipri}
}
\usage{
oc_russie_depenses_militaires
}
\description{
Un tableau de l'évolution des dépenses militaires pour tous
les pays du monde de 2000 à 2021.
}
\keyword{datasets}
